% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncload.R
\name{nc_load}
\alias{nc_load}
\title{Load NetCDF file}
\usage{
nc_load(x)
}
\arguments{
\item{x}{path to NetCDF file}
}
\value{
NULL for success, has the side effect of creating variables in the global environment
}
\description{
Load all the variables from a NetCDF file into the global environment.
}
\details{
See 'nc_read' (not yet implemented) for a less heretical way to do this.
}
\examples{
x <- system.file("extdata", "S2008001.L3m_DAY_CHL_chl_ocx_9km.nc", package = "ncload")
nc_load(x)
}

